+++ packages/framework-playground/src/solid-logo-card.solid.jsx
@@ -1,47 +1,38 @@
 import { createSignal } from "solid-js";
 
 export const SolidLogoCard = () => {
-  const [selectionCount, setSelectionCount] = createSignal(0);
+  const [isSelected, setIsSelected] = createSignal(false);
 
   return (
     <button
       id="solid-logo-button"
       type="button"
       class="framework-logo-button"
-      onClick={() =>
-        setSelectionCount(
-          (previousSelectionCount) => previousSelectionCount + 1,
-        )
-      }
+      onClick={() => setIsSelected((previous) => !previous)}
+      aria-pressed={isSelected()}
     >
       <svg
-        viewBox="0 0 100 100"
+        viewBox="0 0 166 155.3"
         class="framework-logo"
         data-framework-logo="solid"
       >
-        <defs>
-          <linearGradient
-            id="solid-gradient"
-            x1="0%"
-            y1="0%"
-            x2="100%"
-            y2="100%"
-          >
-            <stop offset="0%" stop-color="#76b3e1" />
-            <stop offset="100%" stop-color="#518ac8" />
-          </linearGradient>
-        </defs>
         <path
-          d="M14 24C23 13 37 11 50 15C62 18 74 17 86 9L86 38C76 47 64 50 51 47C39 44 28 45 14 56Z"
-          fill="url(#solid-gradient)"
-        />
-        <path
-          d="M14 56C27 45 39 44 51 47C64 50 76 47 86 38L86 69C76 79 63 83 49 80C36 77 25 79 14 89Z"
+          d="M163 35S110-4 69 5l-3 1c-6 2-11 5-14 9l-2 3-15 26 26 5c11 7 25 10 38 7l46 9 18-30z"
           fill="#76b3e1"
         />
+        <path
+          d="M52 35l-4 1c-17 5-22 21-13 35 10 13 31 20 48 15l62-21S92 26 52 35z"
+          fill="#518ac8"
+        />
+        <path
+          d="M134 80a45 45 0 00-48-15L24 85 4 120l112 19 20-36c4-7 3-15-2-23z"
+          fill="#4377bb"
+        />
+        <path
+          d="M114 115a45 45 0 00-48-15L4 120s53 40 94 30l3-1c17-5 23-21 13-34z"
+          fill="#1a336b"
+        />
       </svg>
-      <span class="framework-logo__name">Solid</span>
-      <span class="framework-logo__count">Selections: {selectionCount()}</span>
     </button>
   );
 };
