+++ packages/framework-playground/src/svelte-logo-card.svelte
@@ -1,28 +1,20 @@
-<script>
-  let selectionCount = 0;
-
-  const incrementSelectionCount = () => {
-    selectionCount += 1;
-  };
-</script>
-
 <button
   id="svelte-logo-button"
   type="button"
   class="framework-logo-button"
-  on:click={incrementSelectionCount}
 >
   <svg
-    viewBox="0 0 100 100"
+    viewBox="0 0 107 128"
     class="framework-logo"
     data-framework-logo="svelte"
   >
-    <circle cx="50" cy="50" r="34" fill="#ff3e00" />
     <path
-      d="M63 31C58 26 51 24 44 25C35 26 29 33 29 41C29 49 35 54 43 56L50 58C54 59 56 61 56 64C56 68 52 71 47 71C42 71 38 69 35 65L28 71C33 78 40 81 49 80C60 79 67 72 67 63C67 55 62 50 53 47L46 45C42 44 40 42 40 39C40 35 43 33 48 33C52 33 56 34 59 38Z"
-      fill="white"
+      d="M94.1566,22.8189c-10.4-14.8851-30.94-19.2971-45.7914-9.8348L22.2825,29.6078A29.9234,29.9234,0,0,0,8.7639,49.6506a31.5136,31.5136,0,0,0,3.1076,20.2318A30.0061,30.0061,0,0,0,7.3953,81.0653a31.8886,31.8886,0,0,0,5.4473,24.1157c10.4022,14.8865,30.9423,19.2966,45.7914,9.8348L84.7167,98.3921A29.9177,29.9177,0,0,0,98.2353,78.3493,31.5263,31.5263,0,0,0,95.13,58.117a30,30,0,0,0,4.4743-11.1824,31.88,31.88,0,0,0-5.4473-24.1157"
+      fill="#ff3e00"
+    />
+    <path
+      d="M45.8171,106.5815A20.7182,20.7182,0,0,1,23.58,98.3389a19.1739,19.1739,0,0,1-3.2766-14.5025,18.1886,18.1886,0,0,1,.6233-2.4357l.4912-1.4978,1.3363.9815a33.6443,33.6443,0,0,0,10.203,5.0978l.9694.2941-.0893.9675a5.8474,5.8474,0,0,0,1.052,3.8781,6.2389,6.2389,0,0,0,6.6952,2.485,5.7449,5.7449,0,0,0,1.6021-.7041L69.27,76.281a5.4306,5.4306,0,0,0,2.4506-3.631,5.7948,5.7948,0,0,0-.9875-4.3712,6.2436,6.2436,0,0,0-6.6978-2.4864,5.7427,5.7427,0,0,0-1.6.7036l-9.9532,6.3449a19.0329,19.0329,0,0,1-5.2965,2.3259,20.7181,20.7181,0,0,1-22.2368-8.2427,19.1725,19.1725,0,0,1-3.2766-14.5024,17.9885,17.9885,0,0,1,8.13-12.0513L55.8833,23.7472a19.0038,19.0038,0,0,1,5.3-2.3287A20.7182,20.7182,0,0,1,83.42,29.6611a19.1739,19.1739,0,0,1,3.2766,14.5025,18.4,18.4,0,0,1-.6233,2.4357l-.4912,1.4978-1.3356-.98a33.6175,33.6175,0,0,0-10.2037-5.1l-.9694-.2942.0893-.9675a5.8588,5.8588,0,0,0-1.052-3.878,6.2389,6.2389,0,0,0-6.6952-2.485,5.7449,5.7449,0,0,0-1.6021.7041L37.73,51.719a5.4218,5.4218,0,0,0-2.4487,3.63,5.7862,5.7862,0,0,0,.9856,4.3717,6.2437,6.2437,0,0,0,6.6978,2.4864,5.7652,5.7652,0,0,0,1.602-.7041l9.9519-6.3425a18.978,18.978,0,0,1,5.2959-2.3278,20.7181,20.7181,0,0,1,22.2368,8.2427,19.1725,19.1725,0,0,1,3.2766,14.5024,17.9977,17.9977,0,0,1-8.13,12.0532L51.1167,104.2528a19.0038,19.0038,0,0,1-5.3,2.3287"
+      fill="#fff"
     />
   </svg>
-  <span class="framework-logo__name">Svelte</span>
-  <span class="framework-logo__count">Selections: {selectionCount}</span>
 </button>
